+++ allennlp/data/samplers/pytorch_samplers.py
@@ -2,6 +2,8 @@ from typing import List, Iterable
 from torch.utils import data
 
 from allennlp.common.registrable import Registrable
+from allennlp.data.samplers.bucket_batch_sampler import BucketBatchSampler
+from allennlp.data.samplers.max_tokens_batch_sampler import MaxTokensBatchSampler
 
 """
 Duplicates of the pytorch Sampler classes. Broadly, these only exist
@@ -11,18 +13,17 @@ pytorch ones.
 """
 
 
-class Sampler(Registrable):
+class PyTorchSampler(Registrable):
     """
     A copy of the pytorch [Sampler](https://pytorch.org/docs/stable/_modules/torch/utils/data/sampler.html)
     which allows us to register it with `Registrable.`
     """
 
     def __iter__(self) -> Iterable[int]:
-
         raise NotImplementedError
 
 
-class BatchSampler(Registrable):
+class PyTorchBatchSampler(Registrable):
     """
     A copy of the pytorch
     [BatchSampler](https://pytorch.org/docs/stable/data.html#torch.utils.data.BatchSampler)
@@ -30,17 +31,16 @@ class BatchSampler(Registrable):
     """
 
     def __iter__(self) -> Iterable[List[int]]:
-
         raise NotImplementedError
 
 
-@Sampler.register("sequential")
-class SequentialSampler(data.SequentialSampler, Sampler):
+@PyTorchSampler.register("sequential")
+class PyTorchSequentialSampler(data.SequentialSampler, PyTorchSampler):
     """
     A registrable version of pytorch's
     [SequentialSampler](https://pytorch.org/docs/stable/data.html#torch.utils.data.SequentialSampler).
 
-    Registered as a `Sampler` with name "sequential".
+    Registered as a `PyTorchSampler` with name "sequential".
 
     In a typical AllenNLP configuration file, `data_source` parameter does not get an entry under
     the "sampler", it gets constructed separately.
@@ -50,24 +50,27 @@ class SequentialSampler(data.SequentialSampler, Sampler):
         super().__init__(data_source)
 
 
-@Sampler.register("random")
-class RandomSampler(data.RandomSampler, Sampler):
+@PyTorchSampler.register("random")
+class PyTorchRandomSampler(data.RandomSampler, PyTorchSampler):
     """
     A registrable version of pytorch's
     [RandomSampler](https://pytorch.org/docs/stable/data.html#torch.utils.data.RandomSampler).
     Samples elements randomly. If without replacement, then sample from a shuffled dataset.
     If with replacement, then user can specify `num_samples` to draw.
 
-    Registered as a `Sampler` with name "random".
+    Registered as a `PyTorchSampler` with name "random".
 
     # Parameters
+
     data_source: `Dataset`, required
         The dataset to sample from.
 
         In a typical AllenNLP configuration file, this parameter does not get an entry under the
         "sampler", it gets constructed separately.
+
     replacement : `bool`, optional (default = `False`)
         Samples are drawn with replacement if `True`.
+
     num_samples: `int` (default = `len(dataset)`)
         The number of samples to draw. This argument
         is supposed to be specified only when `replacement` is ``True``.
@@ -79,16 +82,17 @@ class RandomSampler(data.RandomSampler, Sampler):
         super().__init__(data_source, replacement, num_samples)
 
 
-@Sampler.register("subset_random")
-class SubsetRandomSampler(data.SubsetRandomSampler, Sampler):
+@PyTorchSampler.register("subset_random")
+class PyTorchSubsetRandomSampler(data.SubsetRandomSampler, PyTorchSampler):
     """
     A registrable version of pytorch's
     [SubsetRandomSampler](https://pytorch.org/docs/stable/data.html#torch.utils.data.SubsetRandomSampler).
     Samples elements randomly from a given list of indices, without replacement.
 
-    Registered as a `Sampler` with name "subset_random".
+    Registered as a `PyTorchSampler` with name "subset_random".
 
     # Parameters
+
     indices: `List[int]`
         a sequence of indices to sample from.
     """
@@ -97,14 +101,14 @@ class SubsetRandomSampler(data.SubsetRandomSampler, Sampler):
         super().__init__(indices)
 
 
-@Sampler.register("weighted_random")
-class WeightedRandomSampler(data.WeightedRandomSampler, Sampler):
+@PyTorchSampler.register("weighted_random")
+class PyTorchWeightedRandomSampler(data.WeightedRandomSampler, PyTorchSampler):
     """
     A registrable version of pytorch's
     [WeightedRandomSampler](https://pytorch.org/docs/stable/data.html#torch.utils.data.WeightedRandomSampler).
     Samples elements from `[0,...,len(weights)-1]` with given probabilities (weights).
 
-    Registered as a `Sampler` with name "weighted_random".
+    Registered as a `PyTorchSampler` with name "weighted_random".
 
     # Parameters:
     weights : `List[float]`
@@ -130,17 +134,18 @@ class WeightedRandomSampler(data.WeightedRandomSampler, Sampler):
         super().__init__(weights, num_samples, replacement)
 
 
-@BatchSampler.register("basic")
-class BasicBatchSampler(data.BatchSampler, BatchSampler):
+@PyTorchBatchSampler.register("basic")
+class PyTorchBasicBatchSampler(data.BatchSampler, PyTorchBatchSampler):
     """
     A registrable version of pytorch's
     [BatchSampler](https://pytorch.org/docs/stable/data.html#torch.utils.data.BatchSampler).
     Wraps another sampler to yield a mini-batch of indices.
 
-    Registered as a `BatchSampler` with name "basic".
+    Registered as a `PyTorchBatchSampler` with name "basic".
 
     # Parameters
-    sampler: `Sampler`
+
+    sampler: `PyTorchSampler`
         The base sampler.
     batch_size : `int`
         The size of the batch.
@@ -151,12 +156,63 @@ class BasicBatchSampler(data.BatchSampler, BatchSampler):
     # Examples
 
     ```python
-    >>> list(BatchSampler(SequentialSampler(range(10)), batch_size=3, drop_last=False))
+    list(PyTorchBatchSampler(SequentialSampler(range(10)), batch_size=3, drop_last=False))
     [[0, 1, 2], [3, 4, 5], [6, 7, 8], [9]]
-    >>> list(BatchSampler(SequentialSampler(range(10)), batch_size=3, drop_last=True))
+    list(PyTorchBatchSampler(SequentialSampler(range(10)), batch_size=3, drop_last=True))
     [[0, 1, 2], [3, 4, 5], [6, 7, 8]]
     ```
     """
 
-    def __init__(self, sampler: Sampler, batch_size: int, drop_last: bool):
+    def __init__(self, sampler: PyTorchSampler, batch_size: int, drop_last: bool):
         super().__init__(sampler, batch_size, drop_last)
+
+
+@PyTorchBatchSampler.register("bucket")
+class PyTorchBucketBatchSampler(PyTorchBatchSampler):
+    """
+    A PyTorch-compatible version of `BucketBatchSampler`.
+    """
+
+    def __init__(
+        self,
+        data_source: data.Dataset,
+        batch_size: int,
+        sorting_keys: List[str] = None,
+        padding_noise: float = 0.1,
+        drop_last: bool = False,
+    ) -> None:
+        self.base_sampler = BucketBatchSampler(
+            batch_size, sorting_keys=sorting_keys, padding_noise=padding_noise, drop_last=drop_last
+        )
+        self.data_source = data_source
+
+    def __iter__(self) -> Iterable[List[int]]:
+        return self.base_sampler.get_batch_indices(self.data_source)
+
+    def __len__(self):
+        return self.base_sampler.get_num_batches(self.data_source)
+
+
+@PyTorchBatchSampler.register("max_tokens_sampler")
+class PyTorchMaxTokensBatchSampler(PyTorchBatchSampler):
+    """
+    A PyTorch-compatible version of `MaxTokensBatchSampler`.
+    """
+
+    def __init__(
+        self,
+        data_source: data.Dataset,
+        max_tokens: int,
+        sorting_keys: List[str] = None,
+        padding_noise: float = 0.1,
+    ) -> None:
+        self.base_sampler = MaxTokensBatchSampler(
+            max_tokens, sorting_keys=sorting_keys, padding_noise=padding_noise
+        )
+        self.data_source = data_source
+
+    def __iter__(self) -> Iterable[List[int]]:
+        return self.base_sampler.get_batch_indices(self.data_source)
+
+    def __len__(self):
+        return self.base_sampler.get_num_batches(self.data_source)